requirements.txt
av==10.0.0
torch>=2.0.0
numpy>=1.24.0
transformers>=4.36.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0 